requirements.txt
# MindfulMe - Mental Health Tracker Dependencies

# Core Framework
streamlit>=1.31.0

# Database
sqlalchemy>=2.0.0

# Authentication
bcrypt>=4.1.0

# AI/ML
transformers>=4.36.0
torch>=2.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0

# Data Visualization
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.11.0
